# doc/apifox.openapi.yaml
openapi: 3.0.3
info:
  title: ITSM API Server
  description: |
    由代码路由自动整理的 OpenAPI 3.0 文档，可直接导入 Apifox。

    鉴权说明：
    - 后台接口通常同时要求：
      - JWT：`Authorization: Bearer <token>`
      - 签名：`timestamp`、`nonce`、`signature`，以及 `appkey`（或兼容老头 `client-id`）
    - OpenAPI 对外接口（`/open/*`）要求：
      - `client-id` + `timestamp`、`nonce`、`signature`
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/jira/test:
    post:
      tags: [Jira]
      summary: Jira 测试
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
  /user:
    get:
      tags: [Users]
      summary: 用户列表
      security:
        - bearerAuth: []
          sig_timestamp: []
          sig_nonce: []
          sig_signature: []
          sig_appkey: []
        - bearerAuth: []
          sig_timestamp: []
          sig_nonce: []
          sig_signature: []
          sig_client_id: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    post:
      tags: [Users]
      summary: 创建用户（multipart，可带 files）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /user/{key}:
    get:
      tags: [Users]
      summary: 获取用户信息
      security:
        - bearerAuth: []
          sig_timestamp: []
          sig_nonce: []
          sig_signature: []
          sig_appkey: []
        - bearerAuth: []
          sig_timestamp: []
          sig_nonce: []
          sig_signature: []
          sig_client_id: []
      parameters:
        - in: path
          name: key
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /user/{username}:
    put:
      tags: [Users]
      summary: 更新用户（multipart，可带 files）
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: username
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    delete:
      tags: [Users]
      summary: 删除用户
      security:
        - bearerAuth: []
          sig_timestamp: []
          sig_nonce: []
          sig_signature: []
          sig_appkey: []
        - bearerAuth: []
          sig_timestamp: []
          sig_nonce: []
          sig_signature: []
          sig_client_id: []
      parameters:
        - in: path
          name: username
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /user/{username}/refreshToken:
    put:
      tags: [Users]
      summary: 刷新用户 Token
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: username
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
  /role/{roleId}/users:
    post:
      tags: [Roles]
      summary: 添加用户到角色
      security:
        - bearerAuth: []
          sig_timestamp: []
          sig_nonce: []
          sig_signature: []
          sig_appkey: []
        - bearerAuth: []
          sig_timestamp: []
          sig_nonce: []
          sig_signature: []
          sig_client_id: []
      parameters:
        - in: path
          name: roleId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
  /role/{roleId}/users/{username}:
    delete:
      tags: [Roles]
      summary: 删除角色用户
      security:
        - bearerAuth: []
          sig_timestamp: []
          sig_nonce: []
          sig_signature: []
          sig_appkey: []
        - bearerAuth: []
          sig_timestamp: []
          sig_nonce: []
          sig_signature: []
          sig_client_id: []
      parameters:
        - in: path
          name: roleId
          required: true
          schema: { type: string }
        - in: path
          name: username
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
  /group/{group_name}/users:
    get:
      tags: [Groups]
      summary: 组成员列表
      security:
        - bearerAuth: []
          sig_timestamp: []
          sig_nonce: []
          sig_signature: []
          sig_appkey: []
        - bearerAuth: []
          sig_timestamp: []
          sig_nonce: []
          sig_signature: []
          sig_client_id: []
      parameters:
        - in: path
          name: group_name
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    post:
      tags: [Groups]
      summary: 添加用户到组
      security:
        - bearerAuth: []
          sig_timestamp: []
          sig_nonce: []
          sig_signature: []
          sig_appkey: []
        - bearerAuth: []
          sig_timestamp: []
          sig_nonce: []
          sig_signature: []
          sig_client_id: []
      parameters:
        - in: path
          name: group_name
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /group/{group_name}/users/{username}:
    delete:
      tags: [Groups]
      summary: 从组移除用户
      security:
        - bearerAuth: []
          sig_timestamp: []
          sig_nonce: []
          sig_signature: []
          sig_appkey: []
        - bearerAuth: []
          sig_timestamp: []
          sig_nonce: []
          sig_signature: []
          sig_client_id: []
      parameters:
        - in: path
          name: group_name
          required: true
          schema: { type: string }
        - in: path
          name: username
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sig_timestamp:
      type: apiKey
      in: header
      name: timestamp
      description: 毫秒时间戳（Number）
    sig_nonce:
      type: apiKey
      in: header
      name: nonce
      description: 随机字符串，需避免重复（防重放）
    sig_signature:
      type: apiKey
      in: header
      name: signature
      description: HMAC-SHA256 hex（详见服务端签名算法）
    sig_appkey:
      type: apiKey
      in: header
      name: appkey
      description: App Key（部分接口兼容使用 client-id）
    sig_client_id:
      type: apiKey
      in: header
      name: client-id
      description: OpenAPI Client ID（也可作为 appkey 的兼容头）
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true
        data:
          nullable: true
          description: 具体返回数据，结构随接口而变
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items: {}
            - type: string
            - type: number
            - type: boolean
...